pyproject.toml
[project]
name = "nacosae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "absl-py>=2.3.1",
    "diffusers>=0.35.2",
    "huggingface-hub>=0.36.0",
    "immutabledict>=4.2.2",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pytest-mock>=3.15.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "sphinx>=9.1.0",
    "sphinx-rtd-theme>=3.1.0",
    "streamlit>=1.51.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]

[tool.pytest.ini_options]
pythonpath = "src"
[tool.pylint.format]
max-line-length = 80
[tool.black]
line-length = 80
exclude = "checkpoints"
